+++ demo/midgard-node/src/workers/utils/mpt.ts
@@ -23,7 +23,7 @@ const LEVELDB_ENCODING_OPTS = {
 
 export const makeMpts: Effect.Effect<
   { ledgerTrie: ETH.MerklePatriciaTrie; mempoolTrie: ETH.MerklePatriciaTrie },
-  MptError,
+  SDK.Utils.MptError,
   NodeConfig
 > = Effect.gen(function* () {
   const nodeConfig = yield* NodeConfig;
@@ -51,7 +51,7 @@ export const makeMpts: Effect.Effect<
         useRootPersistence: true,
         valueEncoding: LEVELDB_ENCODING_OPTS.valueEncoding,
       }),
-    catch: (e) => MptError.trieCreate("mempool", e),
+    catch: (e) => SDK.Utils.MptError.trieCreate("mempool", e),
   });
   // Ledger MPT from the other side should use a checkpoint database —
   // its MPT building operations are paired with database ones
@@ -66,7 +66,7 @@ export const makeMpts: Effect.Effect<
         useRootPersistence: true,
         valueEncoding: LEVELDB_ENCODING_OPTS.valueEncoding,
       }),
-    catch: (e) => MptError.trieCreate("ledger", e),
+    catch: (e) => SDK.Utils.MptError.trieCreate("ledger", e),
   });
   const mempoolRootBeforeMempoolTxs = yield* Effect.tryPromise(() =>
     mempoolTrie.get(rootKey),
@@ -96,7 +96,7 @@ export const makeMpts: Effect.Effect<
         );
         yield* Effect.tryPromise({
           try: () => ledgerTrie.batch(ops),
-          catch: (e) => MptError.batch("ledger", e),
+          catch: (e) => SDK.Utils.MptError.batch("ledger", e),
         });
         const rootAfterGenesis = yield* Effect.sync(() => ledgerTrie.root());
         return rootAfterGenesis;
@@ -112,25 +112,31 @@ export const makeMpts: Effect.Effect<
   };
 });
 
-export const deleteMempoolMpt: Effect.Effect<void, MptError, NodeConfig> =
-  Effect.gen(function* () {
-    const config = yield* NodeConfig;
-    yield* Effect.try({
-      try: () =>
-        FS.rmSync(config.MEMPOOL_MPT_DB_PATH, { recursive: true, force: true }),
-      catch: (e) => MptError.trieDelete("mempool", e),
-    });
-  }).pipe(Effect.withLogSpan("Delete mempool MPT"));
+export const deleteMempoolMpt: Effect.Effect<
+  void,
+  SDK.Utils.MptError,
+  NodeConfig
+> = Effect.gen(function* () {
+  const config = yield* NodeConfig;
+  yield* Effect.try({
+    try: () =>
+      FS.rmSync(config.MEMPOOL_MPT_DB_PATH, { recursive: true, force: true }),
+    catch: (e) => SDK.Utils.MptError.trieDelete("mempool", e),
+  });
+}).pipe(Effect.withLogSpan("Delete mempool MPT"));
 
-export const deleteLedgerMpt: Effect.Effect<void, MptError, NodeConfig> =
-  Effect.gen(function* () {
-    const config = yield* NodeConfig;
-    yield* Effect.try({
-      try: () =>
-        FS.rmSync(config.LEDGER_MPT_DB_PATH, { recursive: true, force: true }),
-      catch: (e) => MptError.trieDelete("ledger", e),
-    });
-  }).pipe(Effect.withLogSpan("Delete ledger MPT"));
+export const deleteLedgerMpt: Effect.Effect<
+  void,
+  SDK.Utils.MptError,
+  NodeConfig
+> = Effect.gen(function* () {
+  const config = yield* NodeConfig;
+  yield* Effect.try({
+    try: () =>
+      FS.rmSync(config.LEDGER_MPT_DB_PATH, { recursive: true, force: true }),
+    catch: (e) => SDK.Utils.MptError.trieDelete("ledger", e),
+  });
+}).pipe(Effect.withLogSpan("Delete ledger MPT"));
 
 // Make mempool trie, and fill it with ledger trie with processed mempool txs
 export const processMpts = (
@@ -144,7 +150,7 @@ export const processMpts = (
     mempoolTxHashes: Buffer[];
     sizeOfProcessedTxs: number;
   },
-  MptError | SDK.Utils.CmlUnexpectedError,
+  SDK.Utils.MptError | SDK.Utils.CmlUnexpectedError,
   Database
 > =>
   Effect.gen(function* () {
@@ -190,11 +196,11 @@ export const processMpts = (
       [
         Effect.tryPromise({
           try: () => mempoolTrie.batch(mempoolBatchOps),
-          catch: (e) => MptError.batch("mempool", e),
+          catch: (e) => SDK.Utils.MptError.batch("mempool", e),
         }),
         Effect.tryPromise({
           try: () => ledgerTrie.batch(batchDBOps),
-          catch: (e) => MptError.batch("ledger", e),
+          catch: (e) => SDK.Utils.MptError.batch("ledger", e),
         }),
       ],
       { concurrency: "unbounded" },
@@ -266,35 +272,5 @@ export class LevelDB {
 
   getDatabase() {
     return this._leveldb;
-  }
-}
-
-export class MptError extends Data.TaggedError("MptError")<{
-  readonly message: string;
-  readonly cause?: unknown;
-}> {
-  static get(trie: string, cause?: unknown) {
-    return new MptError({
-      message: `An error occurred on ${trie} trie get operation`,
-      cause,
-    });
-  }
-  static batch(trie: string, cause?: unknown) {
-    return new MptError({
-      message: `An error occurred on ${trie} trie batch operation`,
-      cause,
-    });
-  }
-  static trieDelete(trie: string, cause?: unknown) {
-    return new MptError({
-      message: `An error occurred on whole ${trie} trie delete`,
-      cause,
-    });
-  }
-  static trieCreate(trie: string, cause?: unknown) {
-    return new MptError({
-      message: `An error occurred on ${trie} trie create`,
-      cause,
-    });
   }
 }
