+++ demo/midgard-sdk/src/utils/mpt.ts
@@ -1,18 +1,22 @@
 import { Store, Trie } from "@aiken-lang/merkle-patricia-forestry";
-import { Effect } from "effect";
+import { Data, Effect } from "effect";
 import { inspect } from "node:util";
 
 export const mptFromUTxOs = (
   _spentUtxos: Uint8Array[],
   _producedUtxos: { outputReference: Uint8Array; output: Uint8Array }[],
   ledgerAfterUpdate: { outputReference: Uint8Array; output: Uint8Array }[],
-): Effect.Effect<Trie, Error> =>
+): Effect.Effect<Trie, MptError> =>
   Effect.gen(function* () {
     const store = new Store("utxosStore");
 
     yield* Effect.tryPromise({
       try: () => store.ready(),
-      catch: (e) => new Error(`${e}`),
+      catch: (e) =>
+        new MptError({
+          message: `Failed to initialize UTxO store`,
+          cause: e,
+        }),
     });
 
     /*
@@ -81,7 +85,7 @@ export const mptFromUTxOs = (
       Effect.tryPromise({
         try: () =>
           trie.insert(Buffer.from(outputReference), Buffer.from(output)),
-        catch: (e) => new Error(`${e}`),
+        catch: (e) => MptError.put("UTxO trie", e),
       }),
     );
 
@@ -92,13 +96,17 @@ export const mptFromUTxOs = (
 
 export const mptFromTxs = (
   txs: { txHash: Uint8Array; txCbor: Uint8Array }[],
-): Effect.Effect<Trie, Error> =>
+): Effect.Effect<Trie, MptError> =>
   Effect.gen(function* () {
     const store = new Store("txsStore");
 
     yield* Effect.tryPromise({
       try: () => store.ready(),
-      catch: (e) => new Error(`${e}`),
+      catch: (e) =>
+        new MptError({
+          message: `Failed to initialize txs store`,
+          cause: e,
+        }),
     });
 
     const trie = new Trie(store);
@@ -108,7 +116,7 @@ export const mptFromTxs = (
       ({ txHash, txCbor }) =>
         Effect.tryPromise({
           try: () => trie.insert(Buffer.from(txHash), Buffer.from(txCbor)),
-          catch: (e) => new Error(`${e}`),
+          catch: (e) => MptError.put("txs trie", e),
         }),
       { concurrency: 1 }, // omitting this is equivalent to sequential traversal.
     );
@@ -132,3 +140,39 @@ export const mptFromList = <T>(items: T[]): Effect.Effect<Trie, never, never> =>
     const trie = yield* Effect.promise(() => Trie.fromList(data));
     return trie;
   });
+
+export class MptError extends Data.TaggedError("MptError")<{
+  readonly message: string;
+  readonly cause?: unknown;
+}> {
+  static get(trie: string, cause?: unknown) {
+    return new MptError({
+      message: `An error occurred on ${trie} trie get operation`,
+      cause,
+    });
+  }
+  static put(trie: string, cause?: unknown) {
+    return new MptError({
+      message: `An error occurred on ${trie} trie put operation`,
+      cause,
+    });
+  }
+  static batch(trie: string, cause?: unknown) {
+    return new MptError({
+      message: `An error occurred on ${trie} trie batch operation`,
+      cause,
+    });
+  }
+  static trieDelete(trie: string, cause?: unknown) {
+    return new MptError({
+      message: `An error occurred on whole ${trie} trie delete`,
+      cause,
+    });
+  }
+  static trieCreate(trie: string, cause?: unknown) {
+    return new MptError({
+      message: `An error occurred on ${trie} trie create`,
+      cause,
+    });
+  }
+}
